+++ src/app/decorators.py
@@ -93,7 +93,7 @@ def log_default(
                 res = func(*args, **kwargs)
                 return res
             except Exception as exc:
-                cprint(f"### EXCEPTION in `{func_name}`:\n",
+                cprint(f"### EXCEPTION in `{func_name}`:\n"
                        f"                  {type(exc).__name__}\n"
                        f"                  {str(exc)}", "red", "on_white")