+++ src/organizations/views.py
@@ -10,10 +10,15 @@ from django.contrib.auth.decorators import (
     login_required,
     user_passes_test)
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import (
+    BadRequest,
+    PermissionDenied)
 from django.core.files import File
 from django.core.files.storage import default_storage as storage
 # from django.db.models import Q
-from django.http import HttpResponseRedirect
+from django.http import (
+    HttpResponseForbidden,
+    HttpResponseRedirect)
 from django.shortcuts import (
     get_object_or_404,
     render)
@@ -419,7 +424,7 @@ def organization_details(request, slug=None):
         })
 
 
-@organization_access_check_required
+# @organization_access_check_required
 @log_default(my_logger=logger, cls_or_self=False)
 def organization_staff(request, slug=None):
     """Organization Staff."""
@@ -446,7 +451,7 @@ def organization_staff(request, slug=None):
         })
 
 
-@organization_access_check_required
+# @organization_access_check_required
 @log_default(my_logger=logger, cls_or_self=False)
 def organization_groups(request, slug=None):
     """Organization Groups."""
@@ -484,6 +489,8 @@ def organization_groups(request, slug=None):
 def organization_edit(request, slug=None):
     """Edit Organization."""
     organization = get_object_or_404(Organization, slug=slug)
+    if not organization.is_author(request):
+        raise PermissionDenied
 
     # -------------------------------------------------------------------------
     # --- Prepare Form(s).
@@ -494,18 +501,21 @@ def organization_edit(request, slug=None):
         user=request.user,
         instance=organization)
     aform = AddressForm(
-        request.POST or None, request.FILES or None,
+        request.POST or None,
+        request.FILES or None,
         required=False,
         # required=not request.POST.get("addressless", False),
         instance=organization.address)
 
     formset_phone = PhoneFormSet(
-        request.POST or None, request.FILES or None,
+        request.POST or None,
+        request.FILES or None,
         queryset=Phone.objects.filter(
             content_type=ContentType.objects.get_for_model(organization),
             object_id=organization.id))
     formset_social = SocialLinkFormSet(
-        request.POST or None, request.FILES or None,
+        request.POST or None,
+        request.FILES or None,
         queryset=SocialLink.objects.filter(
             content_type=ContentType.objects.get_for_model(organization),
             object_id=organization.id))
@@ -628,6 +638,8 @@ def organization_populate_newsletter(request, slug=None):
     # --- Initials.
     # -------------------------------------------------------------------------
     organization = get_object_or_404(Organization, slug=slug)
+    if not organization.is_author(request):
+        raise PermissionDenied
 
     # -------------------------------------------------------------------------
     # --- Prepare Form(s).