+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 - Improved SEO Meta ingesting to the Templates.
 - Moved a Media uploading Logic under REST API.
 
+# 0.3.2 (2024-08-09)
+
+### Fix
+
+- Fixed an Access to the Event editing.
+- Fixed an Access to the Organization editing.
+
 # 0.3.1 (2024-07-28)
 
 ### Fix